Rename jsonreturn to jsonresponse.Move db to internal
