+++ api/route/model/createpaste/createpaste.go
@@ -6,20 +6,20 @@ import (
 	"github.com/hashicorp/go-memdb"
 	"github.com/narvikd/ginparser"
 	"net/http"
-	"temppaste/database/paste"
-	"temppaste/internal/jsonreturn"
+	"temppaste/internal/database/paste"
+	"temppaste/internal/jsonresponse"
 )
 
-func Create(ginCtx *gin.Context, trans ut.Translator, DB *memdb.MemDB) (string, *jsonreturn.Model) {
+func Create(ginCtx *gin.Context, trans ut.Translator, DB *memdb.MemDB) (string, *jsonresponse.Model) {
 	model := new(paste.Paste)
 	errParse := ginparser.ParseAndValidate(ginCtx, trans, model)
 	if errParse != nil {
-		return "", jsonreturn.NewModel(http.StatusBadRequest, false, errParse.Error(), "")
+		return "", jsonresponse.NewModel(http.StatusBadRequest, false, errParse.Error(), "")
 	}
 
 	id, err := paste.NewPaste(DB, model)
 	if err != nil {
-		return "", jsonreturn.NewModel(
+		return "", jsonresponse.NewModel(
 			http.StatusInternalServerError, false, "couldn't create paste", "",
 		)
 	}
