+++ api/route/model/getpaste/getpaste.go
@@ -4,14 +4,14 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/hashicorp/go-memdb"
 	"net/http"
-	"temppaste/database/paste"
-	"temppaste/internal/jsonreturn"
+	"temppaste/internal/database/paste"
+	"temppaste/internal/jsonresponse"
 )
 
-func Get(ginCtx *gin.Context, DB *memdb.MemDB) (*paste.Paste, *jsonreturn.Model) {
+func Get(ginCtx *gin.Context, DB *memdb.MemDB) (*paste.Paste, *jsonresponse.Model) {
 	id := ginCtx.Param("id")
 	if id == "" {
-		return nil, jsonreturn.NewModel(
+		return nil, jsonresponse.NewModel(
 			http.StatusBadRequest, false, "id was empty", "",
 		)
 	}
@@ -19,11 +19,11 @@ func Get(ginCtx *gin.Context, DB *memdb.MemDB) (*paste.Paste, *jsonreturn.Model)
 	p, err := paste.GetPaste(DB, id)
 	if err != nil {
 		if err.Error() == "paste not found" {
-			return nil, jsonreturn.NewModel(
+			return nil, jsonresponse.NewModel(
 				http.StatusNotFound, false, err.Error(), "",
 			)
 		}
-		return nil, jsonreturn.NewModel(
+		return nil, jsonresponse.NewModel(
 			http.StatusInternalServerError, false, "couldn't get paste", "",
 		)
 	}
